Scroller stretcher loop
Had to do some wierd font adjustments, but this will work just fine. Also extended the area to 24 chars, as it makes my life a lot easier if it's all on char boundaries, and 16 felt a little cramped. Though it shouldn't be too hard to adjust if need be later.

Also colors are shit right now, but that's a problem that will fix itself when I start on the more fun bits :)