Finish scroller display; only fx left
This was a _bitch_ to get right timing-wise, and in the end I had to add black borders on the top and bottom of the scroller display area, but they actually looked a bit better like this anyways. The nice thing is the rest of this stuff shouldn't affect the display routine - it's set up so we just have to populate the tables it reads from each frame, and we should be able to come up with some cool ways to do that.