Fix small grey pixels at bottom of frame
I had noticed small glitches beneath the frame on a bunch of '64's, but not all of them, and never in the emu. But it looks like in later versions of VICE, these are actually emulated, and it appears to be pretty conservative in the sense that it's always visible when it _could_ happen, unlike on hw where it appears to only happen on some machines.

By chance, because I usually change the border color at the start and end of interrupts to see how long they take, I was able to notice that the position of the grey dots in this demo were always quite close to where the background color change occurred. It would seem that changing the border color during border display can actually cause these graphical glitches. In this demo, they were caused by changing the border color to black at the start of our frame interrupt, which occurs right below the main display area each frame.

Luckily, we never actually change the border color elsewhere in the demo anymore, so fixing this is as easy as moving the background color setup code to the init routine, rather than the frame interrupt. Fixed :)